+++ packages/router/src/bindings/messaging/auctionRequest.ts
@@ -41,9 +41,11 @@ export const auctionRequestBinding = async (
     logger.error("No data in auction request", requestContext, methodContext, err);
     return;
   }
+
   // On every new auction broadcast, route to the new auction handler
   logger.info("Received auction request", requestContext, methodContext);
   const { bid, bidSignature } = await newAuction(data, requestContext);
+
   await messaging.publishAuctionResponse(from, inbox, { bid, bidSignature });
   attemptedAuction.inc({
     sendingAssetId: bid.sendingAssetId,